requirements.txt
flask
flask-cors
requests
pillow
litemapy
nbtlib
gunicorn
numpy
